requirements.txt
# LLM & Agent 프레임워크
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langgraph
langchain-core
langchain-tavily
langchain-community

# 환경 변수
python-dotenv

# 데이터베이스
sqlalchemy
psycopg2-binary

# 웹 서버
fastapi
uvicorn[standard]
websockets

# 벡터 DB (RAG 파이프라인)
chromadb
langchain-chroma

# LangGraph 확장
langgraph-supervisor

# 기타 의존성
typing-extensions
pydantic-settings

# 테스트
pytest
